templates/mcp-server-python/pyproject.toml
[project]
name = "mcp-server-template"
version = "0.1.0"
description = "MCP Server Template with all 10 critical production patterns"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
keywords = ["mcp", "model-context-protocol", "claude", "ai", "tools"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]
auth = [
    # For OAuth 2.1 JWT validation
    "python-jose[cryptography]>=3.3.0",
    # Alternative: PyJWT
    # "PyJWT[crypto]>=2.8.0",
]
http = [
    # For HTTP/SSE transport
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "starlette>=0.27.0",
]

[project.urls]
Homepage = "https://github.com/your-org/mcp-server-template"
Documentation = "https://github.com/your-org/mcp-server-template#readme"
Repository = "https://github.com/your-org/mcp-server-template"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=src --cov-report=term --cov-report=html"

[tool.ruff]
line-length = 100
target-version = "py310"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
